moves applyTheme function to utils
